+++ qa/src/test/fixtures/input/com/example/Model.java
@@ -1,0 +1,8 @@
+package com.example;
+
+import org.realityforge.proton.qa.GenerateType;
+
+@GenerateType
+public class Model
+{
+}